fa-zio-drv.c: move offset and vref to cset
At the moment ZIO is not ready to fully support interleaving. So the
channel attribute are not exportable to userspace through zio_control.
This patch introduce an hack, so the channel attribute are duplicated
and moved to the cset level
NOTE: the correct code is unused, but it still available

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>