fmcadclib.h: add Generic Instance Descriptor
The use of the Generic Instance Descriptor (GID) allows the library to use
a generic pointer during routing operations. When the library routes an
operation to a particular kind of device, the library will do the correct
container_of() on the gid pointer; so, it retrieves the structure associated to
the particular device.

This procedure avoid to use a double pointer during the routing operation.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>