[fix][add] libfmcadc: fix mask control and add a new error
The library was checking the flags field, but it must check the mask field
to verify if there is a configuration to apply. To avoid situations where
the user is convinced to apply a configuration but he forgots to set
the bit mask, now the library return an error if the user tries to configure
with an empty mask.

Thank to "Michel Arruat <michel.arruat@cern.ch>" who found the problem

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>