[bug-fix] fa-zio-drv.c: handle interrupt race condition
There is a little window of time where it happens that the
acquisition_done interrupt come after reading the irq_status but before
the irq acknowledgment. The result is that also acquisition_done is
acked

To solve the problem, the irq handler give the irq acknowledgment only
after it handles all active interrupts. Then, it reads again for active
interrupt and it handles them

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>