kernel: add field-desc.h
This set of macros were proposed by Federico Vaga as part of ZIO, but
I refuse them, for several reasons.  Since this driver uses them
massively and they are not really zio-specific, what was zio-utils.h
is now a drivers-specific header.

This allows to avoid pulling the ZIO commit in master while still
not forcing massive changes to this driver, which is mostly working.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>