kernel: rename hardware-access functions
This has long been a pain for me, and even Federico agrees the initial
choice of names was wrong.  The common_info_get/common_conf_set
function are now renamed to hardware_read/hardware_write, because this
is what actually they are.

People reviewing the code would have a hard time before realizing that
info_get reads a machine register. I hope this solves the issue for
everybody else.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>