kernel: incompatible change: rename all stuff
Unfortunately, fmc-adc is too generic a module name, because we may
expect to have different ADC cards in the same VME rack, at a later
time (and all of them are fmc-adc, in a way).

Also, hardwiring the gateware name as "fmc/spec-fmc-adc-v1.1.bin" is
too specific -- what if 1.2 is a simple fix?

So, the gateware name is a simpler "fmc/adc-100m14b.bin" (which should
be hardware-specific enough while not including a version number.

The module is called fmc-adc-100m14b.ko.

The zio device name is "adc-100m14b" (the fmc- prefix would overflow
the 12-byte maximum device length -- forced by the control structure).
The trigger uses the same name (it's a separate namespace anyways.

This commit trivially updates the manual so all pathnames shown are
correct, leaving a serious manual review to a later time.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>