bugfix: kernel: properly find pre_samples in single-shot
The pre_sample was taken by the trigger-instance values, but there
it was still zero (why?).

So this commit picks it from the current control, where it is properly set.

The bug resulted in PRE+POST samples were taken starting from the
trigger position, resulting in some good (but shifted) samples and a
trailing tail of PRE samples not related to this acquisition (at first
run it is random content of unitialized RAM).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>