lib: turn open/close to new prototypes
This commit makes the library consistent with the official design as
far as open and close are concerned. But buffer selection is still missing.

- open_by_lun() is not there in the low-level implementation, because
the function is expected to be library only, and is expected to fall
back to open, after finding the hardware-specific dev_if from the
administratively-set lun value.

- redundant error checking has been removed. For example, NULL names
are as bad as wild-pointer names, so the program should crash, not
receive an error and manage it. I used this policy in several places.

Other changes brought in by this:

- the board list is now static in boards.c

I'll still move something, at the end: open and close are zio-generic, not
board-specific.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>