api (lib, libtools, doc): : bugfix in prototypes
The prototypes for buffer management so far were wrong.

request_buffer() (which up to now both allocates and fille) is going
to only allocate the buffer, so it doesn't need a timeout.

fill_buffer() on the other hand can have a timeout -- although if
you used ack_poll() or acq_start() with no timetout, the filling is
expected to complete immediately.

The current request_buffer(), which still fills is too,
now uses NULL as a timeout, which isn't an argument any more.
./tools/fald-simple-acq works unchanged, as it already passed NULL
as timeout.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>