lib: support POLLERR when waiting for input samples
Every function that is sable to wait for a timeout now may return -1
with FMCADC_EDISABLED. The new zio commits support that by returning
POLLERR when using poll(2) over a control file descriptor and the
trigger is disabled.  A non-zio driver may support something similar.

Thus, an application may disable the trigger to unlock with error
any thread waiting for an input block.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>