add vmebus.h (copied from svec-sw v2013-12)
Although the file is not yet used (it will be used by svec code),
it is a common file in several projects, so I pick it now to shrink
the diff from the code base currently under test.  The file comes
from current master of svec-sw, which is unchanged since release 2013-12.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>