adc_driver CSET_BUSY flag raised in the IRQ handler
The IRQ handler defers to a workqueue the execution of the DMA.
The workqueue can be delayed (depending of workqueue scheduling policy)
and it may happen that meanwhile a trigger start/stop sequence occurs
to serve a trigger configuration change. Data blocks exist but are
inconsistent with the current acquisition.
To avoid this concurrency issue CSET_BUSY flag is raised
in the IRQ Handler.