adc_driver: channel signal transmisson delay conpensation
Trigger and channel signals are not going through the
same path on the board and trigger is faster.
Trying to sample the trigger itself by connecting
it to a channel, one can see a delay of 20ns between trigger and
its sampling. The compensation is done by delaying the trigger by 2
sample tick (2*10ns)