kernel: fix post-sample=1 acquisition
The ADC start an acquisition only if post-samples > 0. So the acquired
samples are always greater or equal 2. The driver previously tried to
fix this by descrease the number of required post-samples directly in
the hardware. But on user request of post-samples=1 this means request
post-samples=0 to the ADC: so, it will not work.

This patch leave unchanged the logic of the ADC and remove the extra
samples directly from the acquired buffer.

We are aware that this will waste 10ns of acquisition, but the software
implementation to handle this issue is more expensive

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>