kernel: add the possibility to provide custum calibration
Here we do a little hack because I'm already thinking about
a future verison of this driver.

Instead of providing a calibration binary attribute the driver
offers an eeprom_config binary attribute where you can write
data as if it was on the eeprom (calibration data included).
The driver will extract the calibration and configure the ADC.

I'm doing this because we plan to move to platform devices
instead of fmc devices. For this reason the access to the eeprom
will not be easy as today. To begin with we will have a binary
attrbibute where you can write your eeprom content.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>