kernel: do a proper reset of the FMC mezzanine
In gateware version 5 the logic of the FMC reset bit change from active low
to active high: "reset: {1: reset, 0 unreset}".

Here with this patch we do a complete reset cycle of the FMC mezzanine.
The sleeping time between reset and unreset is huge but we do not care much,
this is just the initialization.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>