WIP drv: bugfix on open-drain configuration, voltage range is 1V
We have always been useing an uncalibrated DAC value whenever
we were using 'test data' or the input configuration was in open-drain.

We recently discovered that this is not true, and actually the voltage
range on the input channel is 1V.

For this reason the DAC value still needs to be fixed with eeprom calibration
data.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>
Acked-by: Dimitris Lampridis <Dimitris.Lampridis@cern.ch>