drv: update calibration mechanism
First we change change the values in the GAIN and OFFSET register,
then we apply the configuration by writing 1 in the CALIB_APPLY field.

The patch is more complex than just set one bit becuase it is ready to
implement the calibration thread that will compensate the GAIN factor
based on the current temperature.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>