drv: request DMA channel at probe
On SPEC each ADC has its own DMA channel which can be requested at
probe time. For SVEC, we need to request/release for every
acquisition.

In principle it is also possible for SVEC as well, but this will need
some work on the VME bridge driver to support virtual DMA
channels. And even in that case it could be tricky. Alternativelly we
could implement a share DMA channel, but it is not possible to
implement it in a clean way using the DMA engine framework.

So, this code put down the foundations for the ideal world where we
request the DMA channel once at probe time. But to support the SVEC
case there are *_svec variants in the code.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>