sw: bugfix default setup at init time
The calibration and the default range were not set consistently. The
calibration initialization was running before when the default range was
the result of a memset (0 -> 10V). The code that initialize the voltage
range to 1V happens after but without applying again the calibration
values.

To solve the problem, now the function that sets the voltage range also
sets the calibration values. Ding this allowed to reduce the code in the
voltage range assignment in the ZIO part.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>