Kernel 4.15 compatibility for setup_timer
`setup_timer` is no longer available in kernel 4.15 and onwards,
in favor of `timer_setup`.
Before, we could pass an argument to the callback function as an
unsigned int (generally used with a pointer cast), but now the timer
itself is passed and we retrieve the container struct with a new
macro `from_timer` which is uses `container_of`.
The flags parameter can be set as 0 for the same behavior.

Signed-off-by: Gwenhael GOAVEC <gwenhael.goavec@femto-st.fr>
Signed-off-by: Mathis MARION <mathis.marion@grenoble-inp.org>
Signed-off-by: Federico Vaga <federico.vaga@cern.ch>