sw:drv: fix SVEC completion order
On SVEC carriers we do need to handle shots one by one to be able to
reconfigure the VME DMA window. To handle it we make use of the
completion mechanism: we transfer a shot and we wait its completion,
and then we do the next one.  The completion goes toghether with the
ZIO block, therefore the completion variable is stored togheter with
the block. However, in the fa_dma_complete() code we were cleaning up
the block too soon, so the completion variable is unusable.

Anticipating the completion fixes the problem.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>