kernel: fa-calibration: fix compilation error: 'fa undeclared'
In function 'fa_calib_gain_update()', the first argument ('fa') passed to
'from_timer()' is not declared anywhere, and thus the build throws
compilation error.

Fix this by declaring the variable and then defining it by the value
produced by 'from_timer()'. Then we can use the same variable for the
next steps.

Signed-off-by: Vaibhav Gupta <vaibhav.gupta@cern.ch>
Reviewed-by: Federico Vaga <federico.vaga@cern.ch>