software: kernel: update API for temperature read
In the case of error, the current API just reports via dev_err(), and
return a safe value. This is fine for callibration tools. But other
places, like sysfs, care about the eroor and the error code.

Thus, the API should return the error code in case of error, and the
tools which want a safe value, should do it in their own space.

Signed-off-by: Vaibhav Gupta <vaibhav.gupta@cern.ch>